Update casestudy_t1and2.py
T1 done. We can use this to get T2 done fairly easily by changing up the d_q datastructure. Instead of just getting the driver and passenger that's been waiting the longest amount of time, when we get a passenger, we have to loop through d_q to see which driver is closest (using the euclidean distance fn), and match that driver to the longest waiting passenger. This probably involves changing d_q to a simple list.